+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `smsportal-laravel` will be documented in this file
 
+## 1.1.0 - 2023-01-06
+
+- Added ability to send a single message to multiple numbers at once, including text replacements
+- Added custom exception classes
+- Added unit tests
+
 ## 1.0.0 - 201X-XX-XX
 
 - initial release
+